Module encapsulation: You must export a component from it's module
Angular introduces the concept of module encapsulation in a way similar
to ES modules. It basically means that declarable types - components,
directives, and pipes - can only be used by components declared inside
this module. For example, if I try to use app-customers component from
the customers module inside the App component from the App module, I'll
get an error: Template parse errors: 'app-customers' is not a known
element. This is because there is not 'app-customers' declared in the
App module. If I want to use this component I need to import the module
where this component is defined. This can be done like seen in commit.
The same goes for other declarable types - directives and pipes.